Re-generate SDK using commercetools-api-reference@0f5e29a51620204244388d9590ba04ee8c31014a

+++ src/commercetools/platform/client/images/by_project_key_products_by_id_images_request_builder.py
@@ -34,7 +34,7 @@ class ByProjectKeyProductsByIDImagesRequestBuilder:
         body: typing.BinaryIO,
         *,
         filename: str = None,
-        variant: float = None,
+        variant: int = None,
         sku: str = None,
         staged: bool = None,
         headers: typing.Dict[str, str] = None,
